requirements.txt
pyodbc
ms-fabric-cli
azure-identity